+++ src/main/java/firok/tiths/common/Potions.java
@@ -47,5 +47,6 @@ public final class Potions
 	public static BasePotion eddying = new PotionEddying(); // 旋流
 	public static BasePotion eroded = new PotionEroded(); // 侵蚀
 	public static BasePotion leaves_hiding = new PotionLeavesHiding(); // 蔽叶
+	public static BasePotion illuminating = new PotionIlluminating(); // 灯明
 
 }